updated sol versions + brought in hysi contracts
