+++ packages/hardhat/contracts/core/interfaces/IBeneficiaryVaults.sol
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: MIT
 
-pragma solidity >=0.7.0 <0.8.0;
+pragma solidity ^0.8.0;
 
 interface IBeneficiaryVaults {
   function vaultExists(uint8 vaultId_) external view returns (bool);
