+++ packages/hardhat/contracts/core/interfaces/IStaking.sol
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: MIT
 
-pragma solidity >=0.7.0 <0.8.0;
+pragma solidity ^0.8.0;
 
 interface IStaking {
   function balanceOf(address account) external view returns (uint256);
