+++ packages/hardhat/contracts/core/interfaces/ITokenManager.sol
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: GPL-3.0
 
-pragma solidity >=0.7.0 <0.8.0;
+pragma solidity ^0.8.0;
 
 interface ITokenManager {
   function assignVested(
