+++ packages/hardhat/contracts/core/interfaces/ITreasury.sol
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: MIT
 
-pragma solidity >=0.7.0 <0.8.0;
+pragma solidity ^0.8.0;
 
 interface ITreasury {}
