+++ packages/hardhat/contracts/core/utils/ACLRegistry.sol
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: MIT
 
-pragma solidity >=0.6.0 <0.8.0;
+pragma solidity ^0.8.0;
 
 import "../interfaces/IACLRegistry.sol";
