+++ packages/hardhat/contracts/externals/interfaces/ISetToken.sol
@@ -1,0 +1,191 @@
+/*
+    Copyright 2020 Set Labs Inc.
+
+    Licensed under the Apache License, Version 2.0 (the "License");
+    you may not use this file except in compliance with the License.
+    You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+    Unless required by applicable law or agreed to in writing, software
+    distributed under the License is distributed on an "AS IS" BASIS,
+    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+    See the License for the specific language governing permissions and
+    limitations under the License.
+
+    SPDX-License-Identifier: Apache License, Version 2.0
+*/
+pragma solidity ^0.8.0;
+pragma experimental ABIEncoderV2;
+
+import "@openzeppelin/contracts/token/ERC20/IERC20.sol";
+
+/**
+ * @title ISetToken
+ * @author Set Protocol
+ *
+ * Interface for operating with SetTokens.
+ */
+interface ISetToken is IERC20 {
+  /* ============ Enums ============ */
+
+  enum ModuleState {
+    NONE,
+    PENDING,
+    INITIALIZED
+  }
+
+  /* ============ Structs ============ */
+  /**
+   * The base definition of a SetToken Position
+   *
+   * @param component           Address of token in the Position
+   * @param module              If not in default state, the address of associated module
+   * @param unit                Each unit is the # of components per 10^18 of a SetToken
+   * @param positionState       Position ENUM. Default is 0; External is 1
+   * @param data                Arbitrary data
+   */
+  struct Position {
+    address component;
+    address module;
+    int256 unit;
+    uint8 positionState;
+    bytes data;
+  }
+
+  /**
+   * A struct that stores a component's cash position details and external positions
+   * This data structure allows O(1) access to a component's cash position units and
+   * virtual units.
+   *
+   * @param virtualUnit               Virtual value of a component's DEFAULT position. Stored as virtual for efficiency
+   *                                  updating all units at once via the position multiplier. Virtual units are achieved
+   *                                  by dividing a "real" value by the "positionMultiplier"
+   * @param componentIndex
+   * @param externalPositionModules   List of external modules attached to each external position. Each module
+   *                                  maps to an external position
+   * @param externalPositions         Mapping of module => ExternalPosition struct for a given component
+   */
+  struct ComponentPosition {
+    int256 virtualUnit;
+    address[] externalPositionModules;
+    mapping(address => ExternalPosition) externalPositions;
+  }
+
+  /**
+   * A struct that stores a component's external position details including virtual unit and any
+   * auxiliary data.
+   *
+   * @param virtualUnit       Virtual value of a component's EXTERNAL position.
+   * @param data              Arbitrary data
+   */
+  struct ExternalPosition {
+    int256 virtualUnit;
+    bytes data;
+  }
+
+  /* ============ Functions ============ */
+
+  function addComponent(address _component) external;
+
+  function removeComponent(address _component) external;
+
+  function editDefaultPositionUnit(address _component, int256 _realUnit)
+    external;
+
+  function addExternalPositionModule(
+    address _component,
+    address _positionModule
+  ) external;
+
+  function removeExternalPositionModule(
+    address _component,
+    address _positionModule
+  ) external;
+
+  function editExternalPositionUnit(
+    address _component,
+    address _positionModule,
+    int256 _realUnit
+  ) external;
+
+  function editExternalPositionData(
+    address _component,
+    address _positionModule,
+    bytes calldata _data
+  ) external;
+
+  function invoke(
+    address _target,
+    uint256 _value,
+    bytes calldata _data
+  ) external returns (bytes memory);
+
+  function editPositionMultiplier(int256 _newMultiplier) external;
+
+  function mint(address _account, uint256 _quantity) external;
+
+  function burn(address _account, uint256 _quantity) external;
+
+  function lock() external;
+
+  function unlock() external;
+
+  function addModule(address _module) external;
+
+  function removeModule(address _module) external;
+
+  function initializeModule() external;
+
+  function setManager(address _manager) external;
+
+  function manager() external view returns (address);
+
+  function moduleStates(address _module) external view returns (ModuleState);
+
+  function getModules() external view returns (address[] memory);
+
+  function getDefaultPositionRealUnit(address _component)
+    external
+    view
+    returns (int256);
+
+  function getExternalPositionRealUnit(
+    address _component,
+    address _positionModule
+  ) external view returns (int256);
+
+  function getComponents() external view returns (address[] memory);
+
+  function getExternalPositionModules(address _component)
+    external
+    view
+    returns (address[] memory);
+
+  function getExternalPositionData(address _component, address _positionModule)
+    external
+    view
+    returns (bytes memory);
+
+  function isExternalPositionModule(address _component, address _module)
+    external
+    view
+    returns (bool);
+
+  function isComponent(address _component) external view returns (bool);
+
+  function positionMultiplier() external view returns (int256);
+
+  function getPositions() external view returns (Position[] memory);
+
+  function getTotalComponentRealUnits(address _component)
+    external
+    view
+    returns (int256);
+
+  function isInitializedModule(address _module) external view returns (bool);
+
+  function isPendingModule(address _module) external view returns (bool);
+
+  function isLocked() external view returns (bool);
+}
