+++ packages/hardhat/contracts/externals/interfaces/YearnVault.sol
@@ -1,0 +1,13 @@
+pragma solidity >=0.6.0;
+
+import "@openzeppelin/contracts/token/ERC20/utils/SafeERC20.sol";
+
+interface YearnVault is IERC20 {
+  function token() external view returns (address);
+
+  function deposit(uint256 amount) external;
+
+  function withdraw(uint256 amount) external;
+
+  function pricePerShare() external view returns (uint256);
+}
