+++ packages/hardhat/contracts/mocks/MockERC20.sol
@@ -1,8 +1,8 @@
 // SPDX-License-Identifier: MIT
 
-pragma solidity >=0.6.0 <0.8.0;
+pragma solidity ^0.8.0;
 
-import "@openzeppelin/contracts/token/ERC20/ERC20.sol";
+import "../externals/ERC20.sol";
 
 contract MockERC20 is ERC20 {
   constructor(
