--- packages/hardhat/deploy/05_Hysi.ts
@@ -1,16 +1,0 @@
-import { DeployFunction } from "hardhat-deploy/types";
-import { HardhatRuntimeEnvironment } from "hardhat/types";
-
-const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
-  const { deployments, getNamedAccounts } = hre;
-  const { deploy } = deployments;
-  const { deployer } = await getNamedAccounts();
-
-  await deploy("ContractRegistry", {
-    from: deployer,
-    args: [(await deployments.get("ACLRegistry")).address],
-    log: true,
-    autoMine: true, // speed up deployment on local network (ganache, hardhat), no effect on live networks
-  });
-};
-export default func;
