+++ packages/hardhat/deployments/rinkeby/solcInputs/14ebb96cd28696ce94a006a33973ed49.json
@@ -5,7 +5,7 @@
       "content": "// SPDX-License-Identifier: MIT\n\npragma solidity >=0.6.0 <0.8.0;\n\n/**\n * @dev External interface of AccessControl declared to support ERC165 detection.\n */\ninterface IACLRegistry {\n  /**\n   * @dev Emitted when `newAdminRole` is set as ``role``'s admin role, replacing `previousAdminRole`\n   *\n   * `DEFAULT_ADMIN_ROLE` is the starting admin for all roles, despite\n   * {RoleAdminChanged} not being emitted signaling this.\n   *\n   * _Available since v3.1._\n   */\n  event RoleAdminChanged(\n    bytes32 indexed role,\n    bytes32 indexed previousAdminRole,\n    bytes32 indexed newAdminRole\n  );\n\n  /**\n   * @dev Emitted when `account` is granted `role`.\n   *\n   * `sender` is the account that originated the contract call, an admin role\n   * bearer except when using {AccessControl-_setupRole}.\n   */\n  event RoleGranted(\n    bytes32 indexed role,\n    address indexed account,\n    address indexed sender\n  );\n\n  /**\n   * @dev Emitted when `account` is revoked `role`.\n   *\n   * `sender` is the account that originated the contract call:\n   *   - if using `revokeRole`, it is the admin role bearer\n   *   - if using `renounceRole`, it is the role bearer (i.e. `account`)\n   */\n  event RoleRevoked(\n    bytes32 indexed role,\n    address indexed account,\n    address indexed sender\n  );\n\n  /**\n   * @dev Returns `true` if `account` has been granted `role`.\n   */\n  function hasRole(bytes32 role, address account) external view returns (bool);\n\n  /**\n   * @dev Returns `true` if `account` has been granted `permission`.\n   */\n  function hasPermission(bytes32 permission, address account)\n    external\n    view\n    returns (bool);\n\n  /**\n   * @dev Returns the admin role that controls `role`. See {grantRole} and\n   * {revokeRole}.\n   *\n   * To change a role's admin, use {AccessControl-_setRoleAdmin}.\n   */\n  function getRoleAdmin(bytes32 role) external view returns (bytes32);\n\n  /**\n   * @dev Grants `role` to `account`.\n   *\n   * If `account` had not been already granted `role`, emits a {RoleGranted}\n   * event.\n   *\n   * Requirements:\n   *\n   * - the caller must have ``role``'s admin role.\n   */\n  function grantRole(bytes32 role, address account) external;\n\n  /**\n   * @dev Revokes `role` from `account`.\n   *\n   * If `account` had been granted `role`, emits a {RoleRevoked} event.\n   *\n   * Requirements:\n   *\n   * - the caller must have ``role``'s admin role.\n   */\n  function revokeRole(bytes32 role, address account) external;\n\n  /**\n   * @dev Revokes `role` from the calling account.\n   *\n   * Roles are often managed via {grantRole} and {revokeRole}: this function's\n   * purpose is to provide a mechanism for accounts to lose their privileges\n   * if they are compromised (such as when a trusted device is misplaced).\n   *\n   * If the calling account had been granted `role`, emits a {RoleRevoked}\n   * event.\n   *\n   * Requirements:\n   *\n   * - the caller must be `account`.\n   */\n  function renounceRole(bytes32 role, address account) external;\n\n  function setRoleAdmin(bytes32 role, bytes32 adminRole) external;\n\n  function grantPermission(bytes32 permission, address account) external;\n\n  function revokePermission(bytes32 permission) external;\n\n  function requireApprovedContractOrEOA(address account) external view;\n\n  function requireRole(bytes32 role, address account) external view;\n\n  function requirePermission(bytes32 permission, address account) external view;\n\n  function isRoleAdmin(bytes32 role, address account) external view;\n}\n"
     },
     "contracts/test_helpers/ACLRegistryHelper.sol": {
-      "content": "pragma solidity >=0.7.0 <0.8.0;\n\nimport \"../core/interfaces/IACLRegistry.sol\";\n\ncontract ACLRegistryHelper {\n  IACLRegistry public aclRegistry;\n\n  constructor(IACLRegistry _aclRegistry) {\n    aclRegistry = _aclRegistry;\n  }\n\n  function senderProtected(bytes32 role) public {\n    require(\n      aclRegistry.hasRole(role, msg.sender),\n      \"you dont have the required role\"\n    );\n  }\n}\n"
+      "content": "pragma solidity ^0.8.0;\n\nimport \"../core/interfaces/IACLRegistry.sol\";\n\ncontract ACLRegistryHelper {\n  IACLRegistry public aclRegistry;\n\n  constructor(IACLRegistry _aclRegistry) {\n    aclRegistry = _aclRegistry;\n  }\n\n  function senderProtected(bytes32 role) public {\n    require(\n      aclRegistry.hasRole(role, msg.sender),\n      \"you dont have the required role\"\n    );\n  }\n}\n"
     },
     "contracts/core/utils/ContractRegistry.sol": {
       "content": "// SPDX-License-Identifier: MIT\n\npragma solidity >=0.6.0 <0.8.0;\n\nimport \"../interfaces/IACLRegistry.sol\";\nimport \"../interfaces/IContractRegistry.sol\";\n\n/**\n * @dev This Contract holds reference to all our contracts. Every contract A that needs to interact with another contract B calls this contract\n * to ask for the address of B.\n * This allows us to update addresses in one central point and reduces constructing and management overhead.\n */\ncontract ContractRegistry is IContractRegistry {\n  struct Contract {\n    address contractAddress;\n    bytes32 version;\n  }\n\n  /* ========== STATE VARIABLES ========== */\n\n  IACLRegistry public aclRegistry;\n\n  mapping(bytes32 => Contract) public contracts;\n  bytes32[] public contractNames;\n\n  /* ========== EVENTS ========== */\n\n  event ContractAdded(bytes32 _name, address _address, bytes32 _version);\n  event ContractUpdated(bytes32 _name, address _address, bytes32 _version);\n  event ContractDeleted(bytes32 _name);\n\n  /* ========== CONSTRUCTOR ========== */\n\n  constructor(IACLRegistry _aclRegistry) {\n    aclRegistry = _aclRegistry;\n    contracts[keccak256(\"ACLRegistry\")] = Contract({\n      contractAddress: address(_aclRegistry),\n      version: keccak256(\"1\")\n    });\n    contractNames.push(keccak256(\"ACLRegistry\"));\n  }\n\n  /* ========== VIEW FUNCTIONS ========== */\n\n  function getContractNames() external view returns (bytes32[] memory) {\n    return contractNames;\n  }\n\n  function getContract(bytes32 _name) external view override returns (address) {\n    return contracts[_name].contractAddress;\n  }\n\n  /* ========== MUTATIVE FUNCTIONS ========== */\n\n  function addContract(\n    bytes32 _name,\n    address _address,\n    bytes32 _version\n  ) external {\n    aclRegistry.requireRole(keccak256(\"DAO\"), msg.sender);\n    require(\n      contracts[_name].contractAddress == address(0),\n      \"contract already exists\"\n    );\n    contracts[_name] = Contract({contractAddress: _address, version: _version});\n    contractNames.push(_name);\n    emit ContractAdded(_name, _address, _version);\n  }\n\n  function updateContract(\n    bytes32 _name,\n    address _newAddress,\n    bytes32 _version\n  ) external {\n    aclRegistry.requireRole(keccak256(\"DAO\"), msg.sender);\n    require(\n      contracts[_name].contractAddress != address(0),\n      \"contract doesnt exist\"\n    );\n    contracts[_name] = Contract({\n      contractAddress: _newAddress,\n      version: _version\n    });\n    emit ContractUpdated(_name, _newAddress, _version);\n  }\n\n  function deleteContract(bytes32 _name, uint256 _contractIndex) external {\n    aclRegistry.requireRole(keccak256(\"DAO\"), msg.sender);\n    require(\n      contracts[_name].contractAddress != address(0),\n      \"contract doesnt exist\"\n    );\n    require(\n      contractNames[_contractIndex] == _name,\n      \"this is not the contract you are looking for\"\n    );\n    delete contracts[_name];\n    delete contractNames[_contractIndex];\n    emit ContractDeleted(_name);\n  }\n}\n"
@@ -35,9 +35,7 @@
           "storageLayout",
           "evm.gasEstimates"
         ],
-        "": [
-          "ast"
-        ]
+        "": ["ast"]
       }
     },
     "metadata": {
