+++ packages/hardhat/hardhat.config.ts
@@ -25,7 +25,7 @@ module.exports = {
   solidity: {
     compilers: [
       {
-        version: "0.7.3",
+        version: "0.8.0",
         settings: {
           optimizer: {
             enabled: true,
