+++ packages/hardhat/package.json
@@ -14,7 +14,6 @@
     "test": "yarn hardhat test test/*.test.ts",
     "test:fork": "yarn hardhat test test/fork/*.test.ts",
     "coverage": "yarn hardhat coverage --testfiles 'test/*.test.ts'",
-    "postinstall": "yarn compile",
     "prettify": "yarn lint-staged --allow-empty"
   },
   "lint-staged": {
@@ -23,6 +22,7 @@
   "devDependencies": {
     "@float-capital/solidity-coverage": "^0.7.17",
     "@openzeppelin/test-helpers": "^0.5.13",
+    "@openzeppelin/contracts": "^4.3.2",
     "@types/chai": "^4.2.16",
     "@types/mocha": "^8.2.2",
     "@types/node": "^14.14.41",
@@ -34,7 +34,6 @@
     "@anthonymartin/solidity-docgen": "0.5.13",
     "@chainlink/contracts": "^0.1.7",
     "@nomiclabs/hardhat-waffle": "^2.0.1",
-    "@openzeppelin/contracts": "^3.4.0",
     "@popcorn/utils": "*",
     "@setprotocol/set-protocol-v2": "^0.0.45",
     "@typechain/ethers-v5": "^7.0.0",
